dev-requirements.txt
absl-py==0.7.1
dill==0.2.9
firebase-admin==3.2.1
flake8==3.7.9
funcsigs==1.0.2
gast==0.2.2
gin-config==0.2.1
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigtable==1.2.0
google-cloud-core==1.0.3
google-cloud-firestore==0.31.0
google-cloud-logging==1.14.0
google-cloud-pubsub==1.1.0
google-cloud-storage==1.20.0
google-pasta==0.1.7
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.21.1
gTTS==2.0.4
gTTS-token==1.1.3
gym==0.15.4
h5py==2.9.0
/usr/local/src/jax
/usr/local/src/jaxlib
kubernetes==9.0.0
mne==0.17.1
mock==3.0.5
moviepy==1.0.0
numpy==1.16.2
python-dateutil==2.8.0
PyYAML==5.2
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rfc3339==6.0
scikit-learn==0.22
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
sympy==1.4
tb-nightly==1.14.0a20190603
tensor2tensor==1.15.4
tensorboard==1.15.0
tensorflow-datasets==1.3.2
tensorflow-estimator==1.15.1
tensorflow-gan==2.0.0
tensorflow-hub==0.7.0
tensorflow-metadata==0.15.1
tensorflow-probability==0.7.0
tensorflow-serving-api==1.12.0
tf-estimator-nightly==1.14.0.dev2019060501
trax==1.1.4
yapf==0.29.0
zipp==0.6.0
